Use shared editorVisibilityState
