+++ src/lib/ui/Header.svelte
@@ -2,9 +2,9 @@
   import {enhance} from "$app/forms";
   import {resolve} from "$app/paths";
 
-  import {openEditor} from "$lib/ui/ChannelGroupsEditor.svelte";
   import Center from "$lib/ui/components/Center.svelte";
   import PrimaryButton from "$lib/ui/components/PrimaryButton.svelte";
+  import {showEditor} from "$lib/util/shared.svelte";
 
   type Props = {isSignedIn: boolean};
 
@@ -17,7 +17,7 @@
     {#snippet before()}
       {#if isSignedIn}
         <div class="float-left p-1">
-          <PrimaryButton class="m-1 w-20" onclick={openEditor}>Edit</PrimaryButton>
+          <PrimaryButton class="m-1 w-20" onclick={showEditor}>Edit</PrimaryButton>
         </div>
       {/if}
     {/snippet}
