+++ src/lib/util/shared.svelte.ts
@@ -1,0 +1,13 @@
+let editorVisible: boolean = $state(false);
+
+export function isEditorVisible(): boolean {
+  return editorVisible;
+}
+
+export function showEditor(): void {
+  editorVisible = true;
+}
+
+export function hideEditor(): void {
+  editorVisible = false;
+}
